Move CountryList to its own folder
