+++ src/main.js
@@ -1,7 +1,7 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
 import App from './App.vue';
-import CountryList from './components/CountryList.vue';
+import CountryList from './components/CountryList/CountryList.vue';
 
 Vue.use(VueRouter);
